move genotypes to it's separate file, add methods of importing information about genotypes
