+++ demuxalot/__init__.py
@@ -3,5 +3,6 @@ __author__ = 'Alex Rogozhnikov'
 
 from .utils import BarcodeHandler
 from .snp_counter import count_snps
-from .demux import Demultiplexer, ProbabilisticGenotypes
+from .demux import Demultiplexer
+from .genotypes import ProbabilisticGenotypes
 from .snp_detection import detect_snps_positions
